requirements.txt
streamlit==1.29.0
tensorflow==2.14.0
numpy==1.23.5
pillow==10.0.1